+++ dsp.h
@@ -6,9 +6,9 @@
 #include <inttypes.h>
 
 void dct_quant_block_8x8(int16_t *in_data, int16_t *out_data,
-    uint8_t *quant_tbl);
+    float *quant_tbl);
 
 void dequant_idct_block_8x8(int16_t *in_data, int16_t *out_data,
-    uint8_t *quant_tbl);
+    float *quant_tbl);
 
 #endif  /* C63_DSP_H_ */
